requirements/base.txt
asgiref==3.8.1
attrs==25.3.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.3
django-ckeditor-5==0.2.12
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-import-export==3.3.7
django-jazzmin==3.0.1
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
environ==1.0
et_xmlfile==2.0.0
filelock==3.13.1
flake8==6.1.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupPy==1.18
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
Pillow==10.1.0
pluggy==1.5.0
prometheus_client==0.21.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.3
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.13.2
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
